index: Remove huge Github link
This commit removes the big GitHub block at the bottom of the home page
and instead makes a GitHub button next to the other links in the 'Getting
Started!' section. We make the change both because the block was
excessively big before, and because it will make room for an upcoming
email signup bar.