Style: Fix site layout
There were all sorts of issues with background colors and things
overlapping and doing other ugly things. Since I don't understand web
development very well, the easiest fix was to make the site plain white.

Note that these code changes are not very clean, and a lot of now
outdated code is still in the code base. I consider this a temporary
fix.